+++ WaffleMarket/WaffleMarket/View/CommentTableViewCell.swift
@@ -68,7 +68,7 @@ class CommentTableViewCell: UITableViewCell {
         super.init(coder: coder)
     }
     
-    func setData(isReply: Bool, username: String, profile_image: String, content: String, timestamp: Int){
+    func setData(isReply: Bool, username: String, profile_image: String, content: String, timestamp: Double){
         if isReply {
             containerView.leadingAnchor.constraint(equalTo: self.contentView.leadingAnchor, constant: 40).isActive = true
         } else {
@@ -83,7 +83,7 @@ class CommentTableViewCell: UITableViewCell {
         contentLabel.text = content
         let formatter = DateFormatter()
         formatter.dateFormat = "YYYY/MM/dd hh:mm"
-        timestampLabel.text = formatter.string(from: Date(timeIntervalSince1970: Double(timestamp)))
+        timestampLabel.text = formatter.string(from: Date(timeIntervalSince1970: timestamp))
         contentLabel.heightAnchor.constraint(equalToConstant: content.getEstimatedFrame(with: contentLabel.font).size.height).isActive = true
         
     }
